Intermediate: have Makefiles working to run ut_midcutils.c test.
Still need to finish midcutils unit test and finish starts_with
method in midcutils.  Then have to build the missing files for
the other intended parts of midcutils: a generic stack, a generic
queue and a Ruby-like gsub using the pcre library.
May also want to have the Makefile generate a midcutil library.